vdx_id/requirements.txt
django==2.2.10
djangorestframework==3.11.0

django-allauth==0.41.0
django-celery-beat==2.0.0
django-celery-model==0.2.1
django-celery-results==1.1.2
django-crispy-forms==1.8.1
django-extensions==2.2.6
django-filter ==2.2.0
django-forms-builder==0.14.0
django-fsm==2.7.0
django-fsm-admin==1.2.4
django-fsm-log==1.6.2
django-grappelli==2.13.3
django-json-widget==1.0.0
django-polymorphic==2.1.2
django-redis==4.11.0
django-report-builder==6.3.2
django-rest-swagger==2.2.0
django-simple-history==2.8.0
django-sql-explorer==1.1.3
django-task-api==1.1.6
django-datatables-view==1.19.1
django-notifications-hq==1.6.0
django-admin-tools==0.9.0

aiohttp==2.0.1
boto==2.49.0
bokeh==1.4.0
celery==4.3.1
channels==2.1.2
channels-redis==2.1.1
daphne==2.2.5
dictor==0.1.2
graphviz==0.13.0
ipython==7.11.1
jsonpath-ng==1.4.3
jupyter==1.0.0
lxml==3.5.0 
markdown==3.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.18.1
nxpd==0.2.0
pandas==0.25.3
parse==1.14.0
profilehooks==1.11.0
psycopg2-binary==2.8.4
pydotplus==2.0.2
pymongo==3.9.0
pyquery==1.2.10
pygraphviz==1.5
redis==3.3.11
rules==2.2
textblob==0.15.3
tornado==5.1.1 # Jupyter kernel fixdjango-forms-builder
watchdog==0.9.0
whitenoise==5.0.1
pylint==2.5.3
python-ldap==3.3.0
uWSGI==2.0.17

pytest-django==4.1.0
pytest-cov
